// src/app/core/interface/current-interface.ts
export interface UserCurrent {
    RPEejec: string;
    Nombre: string;
    R3: string;
    IP: string;
    Grupo: string;
    AdminReal: boolean;
    paramInfoPak: string;
    Intro: number;
    Solic: number;
    Busqueda: number;
    Admon: number;
    Directorio: number;
    InfoDoctal: number;
    Chat: number;
    ListaEmail: number;
    Vnc: number;
    VideoConf: number;
    comentarios: string;
    Recursos: number;
    Sugerencias: number;
    xa: number;
    Nivel: number;
    UsuAltas: number;
    UsuModif: number;
    UsuBajas: number;
    BaseCon: number;
    GradosUrgencia: number;
    SolucionesParticulares: number;
    foroAdmin: number;
    foroAcceso: number;
    CoordTiemposAtn: number;
    ConfigAdmOpciones1: number;
    ConfigAdmOpciones2: number;
    ConfigAdmOpciones3: number;
    HayEscalacionAutomatica: number;
    isSupervisor: number;
    isComite: number;
    photo: string;
}